Add complete LeetCode Hot 100 coverage: problems, write-ups and solutions
- Add problems/{NNNN-slug}/README.md for all 100 Hot 100 problems:
  description, examples, constraints, 2-3 solutions compared from
  brute force to optimal, complexity analysis, and common pitfalls
- Add solutions/{NNNN-slug}.cpp for all 100 problems (C++17, LeetCode
  submission format with Solution class only)
- Restructure both READMEs: category-based progress tables linking to
  write-ups and code, remove date column, add Hot 100 badge, update
  directory layout and usage instructions

AI[100%] Human Fixed[0%] Human[0%] AI Adopted[100%]

+++ solutions/0001-two-sum.cpp
@@ -1,11 +1,25 @@
+/**
+ * @file    0001-two-sum.cpp
+ * @brief   1. 两数之和
+ * @link    https://leetcode.cn/problems/two-sum/
+ *
+ * 解法：哈希表
+ * 思路：一遍遍历，哈希表记录「数值 -> 下标」；对当前数先查 target 与其之差是否已出现，再把自己入表。
+ * 时间复杂度：O(n)
+ * 空间复杂度：O(n)
+ */
+
+#include <vector>
+#include <unordered_map>
+
 class Solution {
 public:
     vector<int> twoSum(vector<int>& nums, int target) {
-        unordered_map<int,int>hash;
-        for(int i = 0; i < nums.size(); i++) {
+        unordered_map<int,int> hash;
+        for (int i = 0; i < (int)nums.size(); i++) {
             int now = nums[i];
             int anthor = target - now;
-            if (hash.count(anthor)){
+            if (hash.count(anthor)) {
                 return {i, hash[anthor]};
             }
             hash[now] = i;
